fix: Small fixes from review
