+++ .github/workflows/ci-mcp.yml
@@ -96,10 +96,10 @@ jobs:
             - name: Check generated UI apps are up to date
               run: |
                   cd services/mcp && pnpm run generate:ui-apps
-                  if ! git diff --exit-code -- services/mcp/src/ui-apps/apps/generated/ services/mcp/src/resources/ui-apps.generated.ts; then
+                  if ! git diff --exit-code -- src/ui-apps/apps/generated/ src/resources/ui-apps.generated.ts; then
                     echo ""
                     echo "::warning::The following generated files are out of date:"
-                    git diff --name-only -- services/mcp/src/ui-apps/apps/generated/ services/mcp/src/resources/ui-apps.generated.ts
+                    git diff --name-only -- src/ui-apps/apps/generated/ src/resources/ui-apps.generated.ts
                     echo ""
                     echo "::error::Generated UI app files are out of date. Run 'pnpm --filter=@posthog/mcp run generate:ui-apps' and commit the result."
                     exit 1
