+++ services/mcp/scripts/build-ui-apps.ts
@@ -13,12 +13,12 @@
  */
 import { exec } from 'child_process'
 import { config as dotenvConfig } from 'dotenv'
-import { existsSync, readdirSync, rmSync } from 'fs'
+import { existsSync, rmSync } from 'fs'
 import { join, resolve } from 'path'
 
-const MCP_ROOT_DIR = resolve(__dirname, '..')
-const ROOT_DIR = resolve(MCP_ROOT_DIR, '..', '..')
-const APPS_DIR = resolve(MCP_ROOT_DIR, 'src/ui-apps/apps')
+import { discoverApps, MCP_ROOT_DIR, ROOT_DIR } from './utils'
+
+// Where to store the built files
 const OUT_DIR = resolve(MCP_ROOT_DIR, 'public/ui-apps')
 
 // Load environment variables from .dev.vars (Cloudflare convention)
@@ -27,25 +27,6 @@ if (existsSync(devVarsPath)) {
     const output = dotenvConfig({ path: devVarsPath })
     const loadedKeys = Object.keys(output.parsed || {})
     console.info('📝 Loaded environment from .dev.vars', loadedKeys)
-}
-
-function discoverApps(): string[] {
-    // Top-level custom/manual apps
-    const apps = readdirSync(APPS_DIR)
-        .filter((f) => f.endsWith('.tsx'))
-        .map((f) => f.replace(/\.tsx$/, ''))
-
-    // Generated apps in the generated/ subdirectory
-    const generatedDir = join(APPS_DIR, 'generated')
-    if (existsSync(generatedDir)) {
-        for (const f of readdirSync(generatedDir)) {
-            if (f.endsWith('.tsx')) {
-                apps.push(`generated/${f.replace(/\.tsx$/, '')}`)
-            }
-        }
-    }
-
-    return apps
 }
 
 function buildAppAsync(appName: string): Promise<void> {
